Switch to using nodes and make relevant changes.
Schemas now store maps of nodes instead of maps of validators. This
doesn't matter much now, but in the next few commits nodes (which are
structs) will be expanded to store embedded schemas as well as
validators.

Add two new validators ("additionalItems" and "additionalProperties").
Both used to be handled by their parent validators ("items" and
"properties"), but since they can hold embedded schemas they now need to
be independent validators.

Add NeighborChecker, a new interface for linking nodes. NeighborChecker
can eventually replace SchemaSetter (and will have to for every
validator that can store embedded schemas), but is used now only
for the validators added in this commit, as well as their parent
validators.